requirements.txt
# Web Framework & Server
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
gunicorn>=22,<23
python-multipart>=0.0.9,<1

# Data Validation & Configuration
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-dotenv>=1.0,<2
orjson>=3.10,<4

# HTTP Clients
httpx>=0.27,<1
aiohttp>=3.9,<4.0
requests>=2.31,<3

# Database
SQLAlchemy>=2.0,<3
asyncpg>=0.29 ; platform_system != "Windows"
psycopg2-binary>=2.9

# Azure Services
azure-storage-blob>=12.20,<13
azure-storage-queue==12.9.0
azure-ai-translation-text
azure-ai-translation-document
azure-core==1.30.2
azure-identity==1.17.1

# Authentication & Security
msal>=1.28,<2
PyJWT>=2.8,<3

# Bot Framework
botbuilder-core>=4.14,<5.0
botbuilder-dialogs>=4.14,<5.0
botbuilder-schema>=4.14,<5.0
botbuilder-integration-aiohttp>=4.14,<5.0

# AI & OpenAI
openai>=1.40,<2

# Document Processing
python-docx>=1.1,<2
openpyxl>=3.1,<4
python-pptx>=0.6.23,<1
pdfplumber>=0.11,<1
pypdf
pymupdf>=1.23,<2

# Image Processing
Pillow>=10,<11

# XML Processing
lxml>=5.2,<6

# Data Processing
pandas

# CLI & Utilities
tqdm>=4.66,<5
rich